dataset/requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
argcomplete==3.6.2
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
cftime==1.6.4.post1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
evo==1.31.1
exceptiongroup==1.3.0
ffmpeg-python==0.2.0
filelock==3.13.1
fonttools==4.59.2
fsspec==2024.6.1
future==1.0.0
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hydra-core==1.3.2
idna==3.10
Jinja2==3.1.4
jiter==0.10.0
kiwisolver==1.4.9
lz4==4.4.4
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.6
mdurl==0.1.2
mpmath==1.3.0
mutagen==1.47.0
narwhals==2.4.0
natsort==8.4.0
netCDF4==1.7.2
networkx==3.3
numexpr==2.12.1
numpy==2.2.6
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
omegaconf==2.3.0
openai==1.99.9
opencv-python==4.12.0.88
ordered-set==4.1.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
platformdirs==4.3.8
plotly==6.3.0
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pymap3d==3.2.0
pyparsing==3.2.3
pysofaconventions==0.1.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
rosbags==0.10.11
rsa==4.9.1
rtree==1.4.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
scenedetect==0.6.6
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
stonesoup==1.7
sympy==1.13.3
# torch==2.4.0+cu118
# torchaudio==2.4.0+cu118
torchtyping==0.1.5
# torchvision==0.19.0+cu118
tqdm==4.67.1
triton==3.0.0
typeguard==2.13.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
utm==0.8.1
websockets==15.0.1
yt-dlp==2025.8.22.31729.dev0
zstandard==0.24.0
